delete barber and user updated

+++ backend/src/controllers/barbers/DeleteBarberAndUserAppointmentController.ts
@@ -24,23 +24,27 @@ export class DeleteBarberUserAndAppointmentController {
         throw new AppError('Barber does not exists', 404);
       }
 
-      await prisma.appointment.deleteMany({
-        where: {
-          barberId: id,
-        },
-      });
+      if (barber?.appointment) {
+        await prisma.appointment.deleteMany({
+          where: {
+            barberId: id,
+          },
+        });
+      }
 
-      await prisma.barber.delete({
-        where: {
-          id,
-        },
-      });
+      if (barber?.user) {
+        await prisma.barber.delete({
+          where: {
+            id,
+          },
+        });
 
-      await prisma.user.delete({
-        where: {
-          email: barber.user.email,
-        },
-      });
+        await prisma.user.delete({
+          where: {
+            email: barber.user.email,
+          },
+        });
+      }
 
       return response.json(barber);
     } catch (error) {
